test: split monolithic test class into focused suites
Break the single test class into composition, reference, replacement and
custom-delimiter suites with co-located data, standardize on
Method_State_Expected naming, collapse single-row theories to facts, add a
builder-identity test, assert parameter name and message in option validation,
and drop the unused Moq dependency.
