+++ test/PetToys.TemplatedConfigurationProvider.Tests/PetToys.TemplatedConfigurationProvider.Tests.csproj
@@ -12,7 +12,6 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="Moq" />
     <PackageReference Include="xunit.v3" />
   </ItemGroup>
